Update V7.2
Got rid of read_loop3 and implemented QtTimer approach for instant_measurement3, not to clog up the main thread, dealt with connection error by setting class dependent on QtCore so it is not a plain python class. Adjusted the main update timer to be of self... for python not to bin it.